// packages/trading-core/src/price-service/enigma/types/generated/enigma-price-service.ts
export enum SymbolStatus {
  SETTLING = "SETTLING",
  TRADING = "TRADING",
}
export interface SymbolInfo {
  name: string;
  status: SymbolStatus;
  address: string;
}